simple_python_rsa_aes-master/requirements.txt
pycryptodomex
easygui
pathlib